requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.27
python-telegram-bot==20.6
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.5.2
aiohttp==3.9.1
beautifulsoup4==4.12.2
schedule==1.2.1
pandas==2.1.3
aiogram==3.3.0 